Move non-Host things out of host.ts
